requirements.txt
PyMuPDF
google_cloud_aiplatform==1.101.0
google-cloud-documentai
numpy
opencv_python==4.11.0.86
pandas
protobuf==6.31.1
pypdf==5.8.0
python-dotenv==1.1.1
streamlit==1.46.1
thefuzz==0.22.1
